// src/components/Table/src/types/table.ts
export interface GetColumnsParams {
  ignoreIndex?: boolean;
  ignoreAction?: boolean;
  sort?: boolean;
}
// ...
export type CellFormat = string | ((text: string, record: Recordable, index: number) => string | number) | Map<string | number, any>;

// @ts-ignore
export interface BasicColumn extends ColumnProps<Recordable> {
  children?: BasicColumn[];
  filters?: {
    text: string;
    value: string;
    children?: unknown[] | (((props: Record<string, unknown>) => unknown[]) & (() => unknown[]) & (() => unknown[]));
  }[];

  //
  flag?: 'INDEX' | 'DEFAULT' | 'CHECKBOX' | 'RADIO' | 'ACTION';
  customTitle?: VueNode;

  slots?: Recordable;

  // 自定义header渲染
  customHeaderRender?: (column: BasicColumn) => string | VNodeChild | JSX.Element;

  // Whether to hide the column by default, it can be displayed in the column configuration
  defaultHidden?: boolean;

  // Help text for table column header
  helpMessage?: string | string[];
  /**
   * 格式化
   * 1.表格tag组件
   *   eg: STATUS_OPTIONS = [id: 1, fullName: '待处理', theme:'green', color: '#52C41A', rowKey: 'StatusDesc']
   *       format: 'tag|'+JSON.stringify(STATUS_OPTIONS)
   *   id<number>: 该字段的值，eg status的值、enalbel的值
   *   theme<'green'|'blue'|'yellow'|'red'|'purple'|'gray'>: 内置样式
   *   color<hex>: 颜色值，支持自定义内置颜色以外的色值，该色值为字体颜色，背景色会基于字体颜色淡化。
   *   fullName<string>: 自定义该表格要显示的值
   *   rowKey：显示其他字段的值
   * 2. 日期格式化  eg. 'date| YYYY/MM/DD'
   * 3. CallbackFunction
   */
  format?: CellFormat;

  // Editable
  edit?: boolean;
  editRow?: boolean;
  editable?: boolean;
  editComponent?: ComponentType;
  editComponentProps?:
    | ((opt: { text: string | number | boolean | Recordable; record: Recordable; column: BasicColumn; index: number }) => Recordable)
    | Recordable;
  editRule?: boolean | ((text: string, record: Recordable) => Promise<string>);
  editValueMap?: (value: any) => string;
  onEditRow?: () => void;
  // 权限编码控制是否显示
  auth?: string;
  // 业务控制是否显示
  ifShow?: boolean | ((column: BasicColumn) => boolean);
  // 自定义修改后显示的内容
  editRender?: (opt: {
    text: string | number | boolean | Recordable;
    record: Recordable;
    column: BasicColumn;
    index: number;
  }) => Promise<VNodeChild | JSX.Element>;
  // 动态 Disabled
  editDynamicDisabled?: boolean | ((record: Recordable) => boolean);
  i18nField?: string; // 国际化配置字段
}